refactor: :recycle: move type to entities of each domain
