+++ src/aboutMe/components/certification-card.tsx
@@ -3,7 +3,7 @@ import Card from '@mui/material/Card';
 import CardHeader from '@mui/material/CardHeader';
 import Tooltip from '@mui/material/Tooltip';
 import { FC } from 'react';
-import { Certification } from '../entities/certification';
+import Certification from '../entities/certification';
 
 const CertificationCard: FC<Certification> = (props: Certification) => {
   const { certificationCompany, logo, name, obtentionDate } = props;
