+++ src/aboutMe/entities/certification.tsx
@@ -1,6 +1,8 @@
-export type Certification = {
+type Certification = {
     certificationCompany: string;
     logo: string;
     name: string;
     obtentionDate: string;
 };
+
+export default Certification
