+++ src/aboutMe/entities/hobby.tsx
@@ -1,0 +1,9 @@
+type Hobby = {
+    abbreviation: string;
+    federationLogo: string;
+    name: string;
+    picture: string;
+    sumUp: string;
+};
+
+export default Hobby
