+++ src/aboutMe/entities/study.tsx
@@ -1,0 +1,8 @@
+type Study = {
+    date: string;
+    degree: string;
+    school: string;
+    schoolLogo: string;
+};
+
+export default Study
