+++ src/tabletopGame/entities/tabletopGame.tsx
@@ -1,0 +1,11 @@
+type TabletopGame = {
+    author: string;
+    avatar?: string;
+    date: number;
+    detail: string;
+    picture: string;
+    sumup: string;
+    title: string;
+};
+
+export default TabletopGame
