+++ src/tabletopGame/view/tabletop-game.tsx
@@ -2,7 +2,7 @@ import Grid from '@mui/material/Grid';
 import React from 'react';
 import ReviewCard from '../../shared/components/review-card';
 import { Helmet } from 'react-helmet';
-import { TabletopGame } from '../../types';
+import TabletopGame from '../entities/tabletopGame';
 import { useTranslation } from 'react-i18next';
 
 export default function TabletopGames(): JSX.Element {
